Moved all overall methods to `HypixelBedWars` and created `HypixelBedWarsModes` enum
This commit will create a `HypixelBedWarsModes` enum where each value has a String relating to it's modes JSON key value. They will be used to get information from the API about a certain mode.
